requirements.txt
numpy>=1.25
torch>=2.2
pyyaml>=6.0
tdt
matplotlib>=3.8
scipy>=1.11
h5py>=3.9
scikit-learn>=1.3
